represent monetary amounts as decimal and make optional properties nullable

+++ MailChimp.Net/Models/Promo.cs
@@ -8,7 +8,7 @@ namespace MailChimp.Net.Models
         public string Code { get; set; }
 
         [JsonProperty("amount_discounted")]
-        public double AmountDiscounted { get; set; }
+        public decimal AmountDiscounted { get; set; }
 
         [JsonProperty("type")]
         public string Type { get; set; }
